fix and refactor
